refactor: useOutsideClick 훅으로 모달 바깥 클릭시 모달 꺼짐 추상화
