--- src/components/NavigationBar/ProfileDropdown/Modal/Modal.tsx
@@ -1,3 +1,0 @@
-export default function Modal() {
-  return <></>;
-}
