+++ src/hooks/useOutsideClick.ts
@@ -1,0 +1,25 @@
+import {
+  Dispatch,
+  MutableRefObject,
+  SetStateAction,
+  useEffect,
+  useState,
+} from 'react';
+
+export default function useOutsideClick(
+  ref: MutableRefObject<HTMLDivElement | HTMLUListElement | null>,
+  dropdownOpen: boolean,
+  setDropdownOpen: Dispatch<SetStateAction<boolean>>
+) {
+  const clickModalOutside = (event: any) => {
+    if (dropdownOpen && !ref.current?.contains(event.target)) {
+      setDropdownOpen(false);
+    }
+  };
+  useEffect(() => {
+    document.addEventListener('mousedown', clickModalOutside);
+    return () => {
+      document.removeEventListener('mousedown', clickModalOutside);
+    };
+  });
+}
